Changed [Group1_Tank]_Debugging.R to conform Assignment, Commenting Guidelines, & Object and Methods.
 R code involves the usage of = for assignment within functions, adhering to the specified standard of utilizing = in other assignment applications. Therefore, the code already conforms to the given modified standard.

The given scripy adheres to the modified standard in R, and it includes inline comments following the specified format.

Code is already in compliance with the modified standard in R and does not involve the use of S4 objects or methods. It follows standard S3 object-oriented programming conventions in R.