Changed [Group1_Tank]_Translators.R to conform Assignment, Commenting Guides, Function Docu, & Objects/Methods
The code is already in compliance with the modified standard in R, which involves using " <- " only when assigning functions and using " = " in other assignment applications.

Code is organized as a function create_student_bar_plot that takes a data frame as input and returns the same data frame. Users can use this function by supplying a data frame as input and capturing the result in a variable.

 R script follows this convention, and there's no usage of S4 objects. It uses the standard S3 object-oriented system in R.